requirements.txt
fastapi
uvicorn
httpx
google-genai
python-dotenv
pyzbar
google-cloud-vision
python-multipart
